extract to src/utils as it doesn't assume the DOM
